rename templates and folders so package can be built with poetry masonry
