+++ rs_take_home/utils.py
