+++ setup.cfg
@@ -11,7 +11,7 @@ per-file-ignores =
 # WPS202 Found too many module members
 # WPS442 Found outer scope names shadowing
     tests/*: S101 WPS114 WPS202 WPS442
-    src/utils.py: WPS100
+    rs_take_home/utils.py: WPS100
 
 exclude = docs, .venv, .git
